Overflow was causing me not to find the points at which levels are supposed to be.
Fixed the overflow, now I can read in the level headers correctly, I think.

+++ transAss/readMData11.hpp
@@ -17,7 +17,7 @@ struct fieldRecord {
 };
 
 struct countHeader { //area, teleport, chute header, level offset
-  WORD count;
+  int count;
 };
 
 struct areaRecord {
